refactor: Move search functions and tests to their own files
